Refactor and fix too long page bug

+++ src/helper/get-height.ts
@@ -1,11 +1,12 @@
 export const getHeight = (e: HTMLElement): number => {
   // ignores margin collapse
   const computedStyle = getComputedStyle(e);
+
   return (
     e.clientHeight +
-    parseInt(computedStyle.marginTop) +
-    parseInt(computedStyle.marginBottom) +
-    parseInt(computedStyle.borderTopWidth) +
-    parseInt(computedStyle.borderBottomWidth)
+    parseFloat(computedStyle.marginTop) +
+    parseFloat(computedStyle.marginBottom) +
+    parseFloat(computedStyle.borderTopWidth) +
+    parseFloat(computedStyle.borderBottomWidth)
   );
 };
